refactor(server): move metrics service into package
