+++ packages/doc-server/src/collab.ts
@@ -37,7 +37,7 @@ import {
   recordProjectionRepair,
   recordProjectionReadFallback,
   recordProjectionWipe,
-} from '../../../server/metrics.js';
+} from './metrics.js';
 import { canonicalizeStoredMarks, type StoredMark } from '@proof/core/marks';
 import { refreshSnapshotForSlug } from './snapshot.js';
 import { isShareRole, type ShareRole, type ShareState } from './share-types.js';
