+++ packages/doc-server/src/documents.ts
@@ -49,7 +49,7 @@ import {
   recordRewriteBarrierLatency,
   recordRewriteForceIgnored,
   recordRewriteLiveClientBlock,
-} from '../../../server/metrics.js';
+} from './metrics.js';
 import {
   handleOAuthCallback,
   pollOAuthFlow,
