+++ packages/doc-server/src/snapshot.ts
@@ -4,7 +4,7 @@ import { fileURLToPath } from 'url';
 import { PutObjectCommand, S3Client } from '@aws-sdk/client-s3';
 import { getDocumentBySlug } from '../../../server/db.js';
 import { getCanonicalReadableDocumentSync } from '../../../server/collab.js';
-import { recordSnapshotPublish } from '../../../server/metrics.js';
+import { recordSnapshotPublish } from './metrics.js';
 import { buildSharePreviewModel, renderSharePreviewHtmlPage, resolvePublicOrigin } from './share-preview.js';
 
 const __filename = fileURLToPath(import.meta.url);
